create graph data context
