+++ src/components/Portfolio/PortfolioTabs/PortfolioTabs.tsx
@@ -11,10 +11,7 @@ import TabComponent from '../../Global/TabComponent/TabComponent';
 // import Tokens from '../Tokens/Tokens';
 
 // START: Import Local Files
-import {
-    useAppDispatch,
-    useAppSelector,
-} from '../../../utils/hooks/reduxToolkit';
+import { useAppDispatch } from '../../../utils/hooks/reduxToolkit';
 import { getPositionData } from '../../../App/functions/getPositionData';
 import {
     LimitOrderIF,
@@ -41,6 +38,7 @@ import { LimitOrderServerIF } from '../../../utils/interfaces/LimitOrderIF';
 import { TokenContext } from '../../../contexts/TokenContext';
 import { CachedDataContext } from '../../../contexts/CachedDataContext';
 import { PortfolioTabsPortfolioTabsContainer } from '../../../styled/Components/Portfolio';
+import { GraphDataContext } from '../../../contexts/GraphDataContext';
 
 // interface for React functional component props
 interface propsIF {
@@ -73,18 +71,16 @@ export default function PortfolioTabs(props: propsIF) {
     } = useContext(CrocEnvContext);
     const { lastBlockNumber } = useContext(ChainDataContext);
     const { tokens } = useContext(TokenContext);
+    const { positionsByUser, limitOrdersByUser, changesByUser } =
+        useContext(GraphDataContext);
 
-    const graphData = useAppSelector((state) => state?.graphData);
-    const connectedAccountPositionData =
-        graphData.positionsByUser.positions.filter(
-            (position) => position.chainId === chainId,
-        );
-    const connectedAccountLimitOrderData =
-        graphData.limitOrdersByUser.limitOrders.filter(
-            (position) => position.chainId === chainId,
-        );
-    const connectedAccountTransactionData =
-        graphData.changesByUser.changes.filter((x) => x.chainId === chainId);
+    // TODO: can pull into GraphDataContext
+    const filterFn = <T extends { chainId: string }>(x: T) =>
+        x.chainId === chainId;
+
+    const _positionsByUser = positionsByUser.positions.filter(filterFn);
+    const _txsByUser = changesByUser.changes.filter(filterFn);
+    const _limitsByUser = limitOrdersByUser.limitOrders.filter(filterFn);
 
     const [lookupAccountPositionData, setLookupAccountPositionData] = useState<
         PositionIF[]
@@ -259,15 +255,15 @@ export default function PortfolioTabs(props: propsIF) {
     ]);
 
     const activeAccountPositionData = connectedAccountActive
-        ? connectedAccountPositionData
+        ? _positionsByUser
         : lookupAccountPositionData;
     // eslint-disable-next-line
     const activeAccountLimitOrderData = connectedAccountActive
-        ? connectedAccountLimitOrderData
+        ? _limitsByUser
         : lookupAccountLimitOrderData;
 
     const activeAccountTransactionData = connectedAccountActive
-        ? connectedAccountTransactionData?.filter((tx) => {
+        ? _txsByUser?.filter((tx) => {
               if (tx.changeType !== 'fill' && tx.changeType !== 'cross') {
                   return true;
               } else {
