+++ src/components/Trade/TradeTabs/TradeTabs2.tsx
@@ -8,10 +8,7 @@ import {
     memo,
 } from 'react';
 
-import {
-    useAppDispatch,
-    useAppSelector,
-} from '../../../utils/hooks/reduxToolkit';
+import { useAppSelector } from '../../../utils/hooks/reduxToolkit';
 import useOnClickOutside from '../../../utils/hooks/useOnClickOutside';
 import Transactions from './Transactions/Transactions';
 import Orders from './Orders/Orders';
@@ -24,7 +21,6 @@ import recentTransactionsImage from '../../../assets/images/sidebarImages/recent
 import Ranges from './Ranges/Ranges';
 import TabComponent from '../../Global/TabComponent/TabComponent';
 import PositionsOnlyToggle from './PositionsOnlyToggle/PositionsOnlyToggle';
-import { setChangesByUser } from '../../../utils/state/graphDataSlice';
 import { fetchUserRecentChanges } from '../../../App/functions/fetchUserRecentChanges';
 import Leaderboard from './Ranges/Leaderboard';
 import { DefaultTooltip } from '../../Global/StyledTooltip/StyledTooltip';
@@ -47,6 +43,7 @@ import { FlexContainer } from '../../../styled/Common';
 import { ClearButton } from '../../../styled/Components/TransactionTable';
 import TableInfo from '../TableInfo/TableInfo';
 import { UserDataContext } from '../../../contexts/UserDataContext';
+import { GraphDataContext } from '../../../contexts/GraphDataContext';
 interface propsIF {
     filter: CandleData | undefined;
     setTransactionFilter: Dispatch<SetStateAction<CandleData | undefined>>;
@@ -77,7 +74,7 @@ function TradeTabs2(props: propsIF) {
         server: { isEnabled: isServerEnabled },
     } = useContext(AppStateContext);
     const { chartSettings, tradeTableState } = useContext(ChartContext);
-
+    const { setChangesByUser } = useContext(GraphDataContext);
     const candleTime = chartSettings.candleTime.global;
 
     const {
@@ -107,15 +104,16 @@ function TradeTabs2(props: propsIF) {
         selectedOutsideTab,
     } = useContext(TradeTableContext);
 
-    const graphData = useAppSelector((state) => state?.graphData);
     const tradeData = useAppSelector((state) => state?.tradeData);
     const { isUserConnected, userAddress } = useContext(UserDataContext);
+    const { positionsByUser, limitOrdersByUser, changesByUser } =
+        useContext(GraphDataContext);
 
-    const userChanges = graphData?.changesByUser?.changes;
-    const userLimitOrders = graphData?.limitOrdersByUser?.limitOrders;
-    const userPositions = graphData?.positionsByUser?.positions;
+    const userChanges = changesByUser?.changes;
+    const userLimitOrders = limitOrdersByUser?.limitOrders;
+    const userPositions = positionsByUser?.positions;
 
-    const userPositionsDataReceived = graphData?.positionsByUser.dataReceived;
+    const userPositionsDataReceived = positionsByUser.dataReceived;
 
     const [selectedInsideTab, setSelectedInsideTab] = useState<number>(0);
 
@@ -249,8 +247,6 @@ function TradeTabs2(props: propsIF) {
         diffHashSigPostions(userPositionsMatchingTokenSelection),
     ]);
 
-    const dispatch = useAppDispatch();
-
     useEffect(() => {
         if (
             userAddress &&
@@ -280,12 +276,10 @@ function TradeTabs2(props: propsIF) {
                 })
                     .then((updatedTransactions) => {
                         if (updatedTransactions) {
-                            dispatch(
-                                setChangesByUser({
-                                    dataReceived: true,
-                                    changes: updatedTransactions,
-                                }),
-                            );
+                            setChangesByUser({
+                                dataReceived: true,
+                                changes: updatedTransactions,
+                            });
                         }
                     })
                     .catch(console.error);
